+++ src/common/utils/valid-email.utils.ts
@@ -1,4 +1,4 @@
-import { emailSchema } from '@common/schemas/validation-schemas';
+import { emailSchema } from '@common/schemas/validation.schemas';
 
 export const isValidEmail = (email: string): boolean => {
   return emailSchema.safeParse({ email }).success;
